# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API untuk sistem absensi",
        "title": "Absensi API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3030",
    "basePath": "/api",
    "paths": {
        "/api/attendance/clock-in": {
            "post": {
                "description": "Employee clock in",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attendance"
                ],
                "summary": "Clock in attendance",
                "parameters": [
                    {
                        "description": "Attendance data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Attendance"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Attendance"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/attendance/clock-out/{id}": {
            "put": {
                "description": "Employee clock out",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attendance"
                ],
                "summary": "Clock out attendance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Attendance ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Employee ID",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Attendance"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/attendance/logs": {
            "get": {
                "description": "Get attendance logs with filter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attendance"
                ],
                "summary": "Get attendance logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start Date",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End Date",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Departement ID",
                        "name": "departement_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Employee ID",
                        "name": "employee_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.AttendanceLogResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AttendanceLogResponse": {
            "type": "object",
            "properties": {
                "attendance_id": {
                    "type": "string"
                },
                "clock_in": {
                    "type": "string"
                },
                "clock_out": {
                    "type": "string"
                },
                "date_attendance": {
                    "type": "string"
                },
                "departement_name": {
                    "type": "string"
                },
                "employee_name": {
                    "type": "string"
                },
                "is_late": {
                    "type": "string"
                },
                "is_leave_early": {
                    "type": "string"
                },
                "max_clock_in_time": {
                    "type": "string"
                },
                "max_clock_out_time": {
                    "type": "string"
                },
                "status_clock_in": {
                    "type": "string"
                },
                "status_clock_out": {
                    "type": "string"
                }
            }
        },
...
        "model.Attendance": {
            "type": "object",
            "properties": {
                "attendance_id": {
                    "type": "string"
                },
                "clock_in": {
                    "type": "string"
                },
                "clock_out": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "employee": {
                    "$ref": "#/definitions/model.Employee"
                },
                "employee_id": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.Departement": {
            "type": "object",
            "properties": {
                "departement_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "max_clock_in_time": {
                    "type": "string"
                },
                "max_clock_out_time": {
                    "type": "string"
                }
            }
        },
        "model.Employee": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "departement_id": {
                    "type": "integer"
                },
                "departements": {
                    "$ref": "#/definitions/model.Departement"
                },
                "employee_id": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}